+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7]
+
+### Added
+
+- `generate`: option `--parallel-generation-threads` to specify the number of threads to use for parallel generation.
+  The default is 0, which means no parallel generation.
+
 ## [1.6]
 
 ### Added